pyproject.toml
[tool.poetry]
name = "nflai"
version = "0.1.0"
description = ""
authors = ["shayda-banihashemi <shayda.dev@icloud.com>"]
readme = "README.md"
packages = [{include = "nflai", from = "src"}]

[tool.poetry.dependencies]
python = "^3.10,<3.13"
chromadb = "^0.5.13"
#flask = "^3.1.0"
flask = "2.2.5"
pandas = "^2.2.3"
apache-airflow = "^2.10.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
